LPS-53640 Move default implementation into portal-service
