+++ portal-service/src/com/liferay/portal/model/DefaultLayoutTypeAccessPolicyImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.model.impl;
+package com.liferay.portal.model;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.PortletModeFactory;
@@ -24,11 +24,6 @@ import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.LayoutTypeAccessPolicy;
-import com.liferay.portal.model.LayoutTypePortlet;
-import com.liferay.portal.model.Portlet;
 import com.liferay.portal.security.auth.AuthTokenUtil;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.security.permission.ActionKeys;
@@ -398,8 +393,9 @@ public class DefaultLayoutTypeAccessPolicyImpl
 	}
 
 	private static final boolean _PORTLET_ADD_DEFAULT_RESOURCE_CHECK_ENABLED =
-		GetterUtil.getBoolean(PropsUtil.get(
-			PropsKeys.PORTLET_ADD_DEFAULT_RESOURCE_CHECK_ENABLED));
+		GetterUtil.getBoolean(
+			PropsUtil.get(
+				PropsKeys.PORTLET_ADD_DEFAULT_RESOURCE_CHECK_ENABLED));
 
 	private static final LayoutTypeAccessPolicy _instance =
 		new DefaultLayoutTypeAccessPolicyImpl();
